refactor(webpack): move config files out of root
Move the Webpack config files out of the root directory and into a more
suitable location. This avoids unnecessary clutter of the project root.
