+++ internals/webpack/webpack.config.eslint.js
