+++ internals/webpack/webpack.config.main.prod.js
@@ -2,10 +2,11 @@
  * Webpack config for production electron main process
  */
 
+import path from 'path'
 import webpack from 'webpack'
 import merge from 'webpack-merge'
 import { BundleAnalyzerPlugin } from 'webpack-bundle-analyzer'
-import baseConfig from './webpack.config.base'
+import baseConfig, { rootDir } from './webpack.config.base'
 
 export default merge.smart(baseConfig, {
   devtool: 'source-map',
@@ -14,12 +15,12 @@ export default merge.smart(baseConfig, {
 
   mode: 'production',
 
-  entry: './app/main.dev',
+  entry: path.join(rootDir, 'app', 'main.dev'),
 
   // 'main.js' in root
   output: {
-    path: __dirname,
-    filename: './app/main.prod.js'
+    path: path.join(rootDir, 'app'),
+    filename: 'main.prod.js'
   },
 
   plugins: [
